refactored uikit directory
