+++ frontend/src/shared/components/UIKit/Loader/Loader.scss
